% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MemoryPool.R
\name{MemoryPool}
\alias{MemoryPool}
\title{MemoryPool}
\description{
Manages long-lived memory cells that can be recalled when
distribution shifts are detected. Implements the immunological distinction
between short-lived effector cells and long-lived memory cells.
}
\examples{
# Archive and recall memory cells
data(iris)
X <- as.matrix(iris[, 1:4])
A <- X[sample(150, 10), ]
mp <- MemoryPool$new(archive_threshold = 0.01)
rep <- ImmuneRepertoire$new(A)
mp$archive(rep, X)
mp$size()  # number of archived memories
recalled <- mp$recall(X[1:5, ])
nrow(recalled)  # memories relevant to query

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{memory_cells}}{Numeric matrix of archived memory antibodies.}

\item{\code{memory_metadata}}{Data frame of metadata for memory cells.}

\item{\code{archive_threshold}}{Affinity threshold for archiving (only high-quality
antibodies become memory).}

\item{\code{max_memory}}{Maximum number of memory cells to store.}

\item{\code{recall_threshold}}{Threshold for triggering memory recall.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MemoryPool-new}{\code{MemoryPool$new()}}
\item \href{#method-MemoryPool-archive}{\code{MemoryPool$archive()}}
\item \href{#method-MemoryPool-recall}{\code{MemoryPool$recall()}}
\item \href{#method-MemoryPool-size}{\code{MemoryPool$size()}}
\item \href{#method-MemoryPool-print}{\code{MemoryPool$print()}}
\item \href{#method-MemoryPool-clone}{\code{MemoryPool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryPool-new"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryPool-new}{}}}
\subsection{Method \code{new()}}{
Create a new MemoryPool.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryPool$new(
  archive_threshold = 0.5,
  max_memory = 100,
  recall_threshold = 0.3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive_threshold}}{Numeric. Minimum average affinity to archive.}

\item{\code{max_memory}}{Integer. Maximum memory cells.}

\item{\code{recall_threshold}}{Numeric. Minimum affinity to recall a memory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryPool-archive"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryPool-archive}{}}}
\subsection{Method \code{archive()}}{
Archive high-performing antibodies as memory cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryPool$archive(
  repertoire,
  X,
  affinityFunc = "gaussian",
  affinityParams = list(alpha = 1, c = 1, p = 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repertoire}}{An \code{\link{ImmuneRepertoire}}.}

\item{\code{X}}{Training data matrix.}

\item{\code{affinityFunc}}{Character. Affinity function.}

\item{\code{affinityParams}}{List. Affinity parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Integer. Number of new memory cells archived.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryPool-recall"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryPool-recall}{}}}
\subsection{Method \code{recall()}}{
Recall memory cells relevant to current data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryPool$recall(
  X,
  affinityFunc = "gaussian",
  affinityParams = list(alpha = 1, c = 1, p = 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data matrix to match against memory.}

\item{\code{affinityFunc}}{Character. Affinity function.}

\item{\code{affinityParams}}{List. Affinity parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numeric matrix of recalled memory cells (may be empty).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryPool-size"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryPool-size}{}}}
\subsection{Method \code{size()}}{
Get current memory pool size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryPool$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryPool-print"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryPool-print}{}}}
\subsection{Method \code{print()}}{
Print summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryPool$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryPool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryPool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryPool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
