% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-coalesce operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to test.}

\item{y}{Default value if \code{x} is NULL.}
}
\value{
\code{x} if not NULL, otherwise \code{y}.
}
\description{
Returns \code{x} if it is not \code{NULL}, otherwise returns \code{y}.
Used internally by bHIVE R6 classes for parameter defaults.
}
\keyword{internal}
